app/requirements.txt
Flask
kafka-python
pykafka